fix: cast transaction status to i16 in change_transaction_status and add debug log for cancellation

+++ src/db/transaction_repository.rs
@@ -105,7 +105,7 @@ impl TransactionRepository {
             SET status = ?
             WHERE id = ?
             "#,
-            status,
+            status as i16,
             transaction_id
         )
         .execute(&self.pool)
